+++ fakelab.config.ts
@@ -1,23 +1,8 @@
-import { defineConfig } from "fakelab";
+import { defineConfig } from "./src/main";
 
 export default defineConfig({
   sourcePath: ["./fixtures"],
-  enabled: false,
   server: { port: 50001, includeSnapshots: true },
-  snapshot: {
-    enabled: true,
-    sources: [
-      {
-        name: "Todo",
-        url: "https://jsonplaceholder.typicode.com/todos",
-      },
-      {
-        name: "Post",
-        url: "https://jsonplaceholder.typicode.com/posts",
-      },
-    ],
-  },
   database: { enabled: true },
-  graphQL: { enabled: true },
-  network: { delay: [500, 1500] },
+  headless: false,
 });