+++ fixtures/user.ts
@@ -1,13 +1,33 @@
 export interface User {
+  /**
+   * @faker string.ulid
+   */
   id: string;
+  /**
+   * @faker person.fullName
+   */
   name: string;
+  /**
+   * @faker location.streetAddress
+   */
+  address: string;
+  /**
+   * @faker phone.number
+   */
+  phone: string;
+  /**
+   * @faker number.int({min:10,max:80})
+   */
   age: number;
-  active: boolean;
+
+  /**
+   * @faker person.fullName
+   */
+  tags: string[];
+
+  profile: Profile;
 }
 
-export interface Post {
-  id: string;
-  title: string;
-  description: string;
-  image: string;
+export interface Profile {
+  name: string;
 }