+++ src/cli.ts
@@ -18,6 +18,8 @@ program
   .option("-x, --pathPrefix <char>", "server url path prefix")
   .option("-p, --port <number>", "server port number", parseInt)
   .option("-l, --locale <char>", "faker custom locale")
+  .option("-h, --headless", "headless mode")
+  .option("-c, --ts-config <char>", "ts config path")
   .option("-f, --fresh-snapshots", "capture or refresh all snapshots")
   .action(async (options) => {
     const snapshot = await Snapshot.prepare(options);