+++ src/database/index.ts
@@ -1,20 +1,20 @@
 import fs from "fs-extra";
 import path from "node:path";
-import type { DatabaseOptions } from "./types";
-import { CWD } from "./file";
-import { Logger } from "./logger";
-import type { Config } from "./config/conf";
+import type { DatabaseOptions } from "../types";
+import { CWD } from "../file";
+import type { Config } from "../config/conf";
+import { Logger } from "../logger";
 
 export class Database {
   private options: Required<DatabaseOptions>;
-  readonly DATABASE_DIR = path.resolve(CWD, ".fakelab/db");
+  static readonly DATABASE_DIR = path.resolve(CWD, ".fakelab/db");
 
   private constructor(private readonly config: Config) {
     this.enabled = this.enabled.bind(this);
 
     this.options = this.config.options.database();
 
-    if (!this.options.enabled) fs.rmSync(this.DATABASE_DIR, { force: true, recursive: true });
+    if (!this.options.enabled) fs.rmSync(Database.DATABASE_DIR, { force: true, recursive: true });
   }
 
   static register(config: Config) {
@@ -32,7 +32,7 @@ export class Database {
         return;
       }
       try {
-        await fs.ensureDir(this.DATABASE_DIR);
+        await fs.ensureDir(Database.DATABASE_DIR);
 
         await this.modifyGitignoreFile(".fakelab/*");
       } catch (error) {